Make gns3server.appliances a package

+++ gns3server/controller/__init__.py
@@ -41,7 +41,6 @@ from ..version import __version__
 from .topology import load_topology
 from .gns3vm import GNS3VM
 from .gns3vm.gns3_vm_error import GNS3VMError
-from gns3server import configs as gns3_configs
 
 import logging
 log = logging.getLogger(__name__)
@@ -289,12 +288,11 @@ class Controller:
                     if not os.path.exists(os.path.join(dst_path, filename)):
                         shutil.copy(os.path.join(resource_path, filename), os.path.join(dst_path, filename))
             else:
-                for entry in importlib_resources.files(gns3_configs).iterdir():
-                    if entry.is_file():
-                        full_path = os.path.join(dst_path, entry.name)
-                        if not os.path.exists(full_path):
-                            log.debug(f"Installing base config file {entry.name} to {full_path}")
-                            shutil.copy(str(entry), os.path.join(dst_path, entry.name))
+                for entry in importlib_resources.files('gns3server.configs').iterdir():
+                    full_path = os.path.join(dst_path, entry.name)
+                    if entry.is_file() and not os.path.exists(full_path):
+                        log.debug(f"Installing base config file {entry.name} to {full_path}")
+                        shutil.copy(str(entry), os.path.join(dst_path, entry.name))
         except OSError as e:
             log.error(f"Could not install base config files to {dst_path}: {e}")
